Hidding Laguerre and turning off tests
